Make POS default discount a whole percentage and keep it after clearing
The public POS views can only express whole-percent discounts: the cart
input is step="1" and the keypad enters the discount with divisibility 0.
A fractional default therefore pre-filled the cart with a value that fails
the input's step validation, which blocks the form submit and leaves the
customer unable to pay.

Clearing the cart also reset the discount to null instead of the store's
default, so on a terminal that stays open the default only survived the
first sale.

Cover both in the POS Playwright tests, and restore the ShowDiscount
default assertion in the Greenfield test by moving the discount coverage
into its own app.

+++ BTCPayServer.Client/Models/PointOfSaleAppData.cs
@@ -12,7 +12,7 @@ public abstract class PointOfSaleBaseData : AppBaseData
     public bool? ShowItems { get; set; }
     public bool? ShowCustomAmount { get; set; }
     public bool? ShowDiscount { get; set; }
-    public decimal? DefaultDiscount { get; set; }
+    public int? DefaultDiscount { get; set; }
     public bool? ShowSearch { get; set; }
     public bool? ShowCategories { get; set; }
     public bool? EnableTips { get; set; }
