+++ BTCPayServer/Services/Apps/PointOfSaleSettings.cs
@@ -93,7 +93,7 @@ namespace BTCPayServer.Services.Apps
         public bool ShowItems { get; set; }
         public bool ShowCustomAmount { get; set; }
         public bool ShowDiscount { get; set; }
-        public decimal? DefaultDiscount { get; set; }
+        public int? DefaultDiscount { get; set; }
         public bool ShowSearch { get; set; }
         public bool ShowCategories { get; set; }
         public bool EnableTips { get; set; }
